Diverse oppdateringer

+++ Web2Prosjektoppgave.api/Controllers/BlogPostController.cs
@@ -5,6 +5,7 @@ using System.Security.Claims;
 using System.Text.RegularExpressions;
 using Web2Prosjektoppgave.api.Models.Entities;
 using Web2Prosjektoppgave.api.Models.Interfaces;
+using Web2Prosjektoppgave.api.Security;
 using Web2Prosjektoppgave.shared.ViewModels.BlogPost;
 using Web2Prosjektoppgave.shared.ViewModels.Comment;
 
@@ -78,7 +79,7 @@ namespace Web2Prosjektoppgave.api.Controllers
         }
 
         // Post
-        //[Authorize]
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> Create(BlogPostCreateForm blogPostForm)
         {
@@ -94,25 +95,14 @@ namespace Web2Prosjektoppgave.api.Controllers
                 return NotFound();
             }
 
-            //var requirement = new UserRequirement(blog.CreatedById);
-            //// Check if the current user has the same ID as the one being requested
-            //var authorizationResult = await _authorizationService.AuthorizeAsync(User, null, requirement);
+            var requirement = new UserRequirement(blog.CreatedById);
+            // Check if the current user has the same ID as the one being requested
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, null, requirement);
 
-            //if (!authorizationResult.Succeeded)
-            //{
-            //    return new ForbidResult();
-            //}
-
-            //var tagStrings = blogPost.Tags.Split(",");
-            //var blogPostTags = new List<BlogPostTag>();
-            //foreach (var tag in tagStrings)
-            //{
-            //    blogPostTags.Add(new BlogPostTag()
-            //    {
-            //        Name = tag
-            //    });
-            //}
-
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
 
             var tagList = new List<BlogPostTag>();
             var regex = new Regex(@"#\w+");
@@ -135,10 +125,8 @@ namespace Web2Prosjektoppgave.api.Controllers
 
             var postModel = new BlogPost()
             {
-                CreatedById = 1,
-                ModifiedById = 1,
-                //CreatedById = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value),
-                //ModifiedById = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value),
+                CreatedById = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value),
+                ModifiedById = int.Parse(User.FindFirst(ClaimTypes.NameIdentifier).Value),
                 Title = blogPostForm.Title,
                 Content = blogPostForm.Content,
                 BlogId = blogPostForm.BlogId,
@@ -167,14 +155,14 @@ namespace Web2Prosjektoppgave.api.Controllers
                 return NotFound();
             }
 
-            //var requirement = new UserRequirement(updatePost.CreatedById);
-            //// Check if the current user has the same ID as the one being requested
-            //var authorizationResult = await _authorizationService.AuthorizeAsync(User, null, requirement);
+            var requirement = new UserRequirement(updatePost.CreatedById);
+            // Check if the current user has the same ID as the one being requested
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, null, requirement);
 
-            //if (!authorizationResult.Succeeded)
-            //{
-            //    return new ForbidResult();
-            //}
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
 
             var tagList = new List<BlogPostTag>();
             var regex = new Regex(@"#\w+");
@@ -221,14 +209,14 @@ namespace Web2Prosjektoppgave.api.Controllers
                 return NotFound();
             }
 
-            //var requirement = new UserRequirement(post.CreatedById);
-            //// Check if the current user has the same ID as the one being requested
-            //var authorizationResult = await _authorizationService.AuthorizeAsync(User, null, requirement);
+            var requirement = new UserRequirement(blogPost.CreatedById);
+            // Check if the current user has the same ID as the one being requested
+            var authorizationResult = await _authorizationService.AuthorizeAsync(User, null, requirement);
 
-            //if (!authorizationResult.Succeeded)
-            //{
-            //    return new ForbidResult();
-            //}
+            if (!authorizationResult.Succeeded)
+            {
+                return new ForbidResult();
+            }
 
             await _blogPostRepository.Delete(blogPost);
 
@@ -237,23 +225,38 @@ namespace Web2Prosjektoppgave.api.Controllers
 
         //[Authorize]
         [HttpGet]
-        [Route("Search")]
-        public async Task<IActionResult> Search(string phrase)
+        [Route("Search/{blogId}")]
+        public async Task<IActionResult> Search(string phrase, int blogId)
         {
-            var blogPosts = await _blogPostRepository.Search(phrase);
-            var blogIdList = new List<int>();
+            var blogPosts = await _blogPostRepository.Search(phrase, blogId);
+            var listOfBlogPosts = new List<BlogPostItemView>();
 
-            foreach (var blogpost in blogPosts)
+            foreach (var blogPost in blogPosts)
             {
-                blogIdList.Add(blogpost.BlogId);
+                var commentRequest = await _commentRepository.GetAllByBlogPostId(blogPost.Id);
+
+                listOfBlogPosts.Add(new BlogPostItemView()
+                {
+                    Id = blogPost.Id,
+                    CreatedAt = blogPost.CreatedAt,
+                    CreatedByUserName = blogPost.CreatedBy.UserName,
+                    ModifiedAt = blogPost.ModifiedAt,
+                    ModifiedByUserName = blogPost.ModifiedBy.UserName,
+                    Title = blogPost.Title,
+                    Content = blogPost.Content,
+                    BlogComments = commentRequest.Select(comment => new CommentItemView()
+                    {
+                        Id = comment.Id,
+                        CreatedAt = comment.CreatedAt,
+                        CreatedByUserName = comment.CreatedBy.UserName,
+                        ModifiedAt = comment.ModifiedAt,
+                        ModifiedByUserName = comment.ModifiedBy.UserName,
+                        Content = comment.Content,
+                    }).ToList(),
+            });
             }
 
-            var searchedBlogIds = new BlogPostSearch()
-            {
-                BlogIds = blogIdList
-            };
-
-            return Ok(searchedBlogIds);
+            return Ok(listOfBlogPosts);
         }
     }
 }
